[ref] move to pkg
