+++ src/main/java/org/ieeervce/api/siterearnouveau/dto/execom/ExecomMemberDTO.java
@@ -1,4 +1,4 @@
-package org.ieeervce.api.siterearnouveau.dto;
+package org.ieeervce.api.siterearnouveau.dto.execom;
 
 public class ExecomMemberDTO {
     String name;
