+++ src/main/java/org/ieeervce/api/siterearnouveau/dto/execom/ExecomYearListDTO.java
@@ -1,4 +1,4 @@
-package org.ieeervce.api.siterearnouveau.dto;
+package org.ieeervce.api.siterearnouveau.dto.execom;
 
 import java.util.HashMap;
 import java.util.Map;
